remake project structure and add meta files
