+++ LICENSE.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 62877d9ad8ac644c7a62787e1d12df9a
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
