+++ README.md.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: e8d0526a79d1f4910935f0bbbc75d11a
+TextScriptImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
