+++ Relocator.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 9c9bfcceade2d429cb578912b5866b82
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
